+++ client/app/reader/reducer.js
@@ -104,7 +104,7 @@ export const initialState = {
   extractedText: {}
 };
 
-const readerReducer = (state = initialState, action = {}) => {
+export const reducer = (state = initialState, action = {}) => {
   let allTags;
   let uniqueTags;
 
@@ -520,6 +520,6 @@ const readerReducer = (state = initialState, action = {}) => {
 };
 
 export default timeFunction(
-  readerReducer,
+  reducer,
   (timeLabel, state, action) => `Action ${action.type} reducer time: ${timeLabel}`
 );
