+++ client/test/karma/reader/selectors-test.js
@@ -10,10 +10,12 @@ describe('Reader utils', () => {
           3: {},
           5: {}
         },
-        ui: {
-          filteredDocIds: [1, 3, 5],
-          docFilterCriteria: {
-            searchQuery: ''
+        readerReducer: {
+          ui: {
+            filteredDocIds: [1, 3, 5],
+            docFilterCriteria: {
+              searchQuery: ''
+            }
           }
         }
       };
@@ -28,10 +30,12 @@ describe('Reader utils', () => {
           3: {},
           5: {}
         },
-        ui: {
-          filteredDocIds: [1, 3, 5],
-          docFilterCriteria: {
-            searchQuery: 'something that matches all docs'
+        readerReducer: {
+          ui: {
+            filteredDocIds: [1, 3, 5],
+            docFilterCriteria: {
+              searchQuery: 'something that matches all docs'
+            }
           }
         }
       };
@@ -46,12 +50,14 @@ describe('Reader utils', () => {
           3: {},
           5: {}
         },
-        ui: {
-          filteredDocIds: [1, 3, 5],
-          docFilterCriteria: {
-            searchQuery: '',
-            category: {
-              procedural: true
+        readerReducer: {
+          ui: {
+            filteredDocIds: [1, 3, 5],
+            docFilterCriteria: {
+              searchQuery: '',
+              category: {
+                procedural: true
+              }
             }
           }
         }
@@ -67,12 +73,14 @@ describe('Reader utils', () => {
           3: {},
           5: {}
         },
-        ui: {
-          filteredDocIds: [1, 3, 5],
-          docFilterCriteria: {
-            searchQuery: '',
-            tag: {
-              'some tag': true
+        readerReducer: {
+          ui: {
+            filteredDocIds: [1, 3, 5],
+            docFilterCriteria: {
+              searchQuery: '',
+              tag: {
+                'some tag': true
+              }
             }
           }
         }
@@ -86,58 +94,60 @@ describe('Reader utils', () => {
     it('gets annotations', () => {
       const documentId = 700;
       const state = {
-        editingAnnotations: {
-          100: {
-            id: 100,
-            comment: 'edited',
-            documentId
+        annotationLayer: {
+          editingAnnotations: {
+            100: {
+              id: 100,
+              comment: 'edited',
+              documentId
+            },
+            300: {
+              id: 300,
+              comment: 'wrong doc',
+              documentId: 800
+            }
           },
-          300: {
-            id: 300,
-            comment: 'wrong doc',
-            documentId: 800
-          }
-        },
-        annotations: {
-          100: {
-            id: 100,
-            comment: 'original',
-            documentId
+          annotations: {
+            100: {
+              id: 100,
+              comment: 'original',
+              documentId
+            },
+            200: {
+              id: 200,
+              comment: 'original 2',
+              documentId
+            },
+            250: {
+              id: 250,
+              comment: 'original 3',
+              documentId
+            },
+            270: {
+              id: 270,
+              comment: 'should be deleted',
+              documentId,
+              pendingDeletion: true
+            },
+            300: {
+              id: 300,
+              comment: 'different doc',
+              documentId: 800
+            }
           },
-          200: {
-            id: 200,
-            comment: 'original 2',
-            documentId
+          pendingEditingAnnotations: {
+            250: {
+              id: 250,
+              comment: 'pending edit',
+              documentId
+            }
           },
-          250: {
-            id: 250,
-            comment: 'original 3',
-            documentId
-          },
-          270: {
-            id: 270,
-            comment: 'should be deleted',
-            documentId,
-            pendingDeletion: true
-          },
-          300: {
-            id: 300,
-            comment: 'different doc',
-            documentId: 800
-          }
-        },
-        pendingEditingAnnotations: {
-          250: {
-            id: 250,
-            comment: 'pending edit',
-            documentId
-          }
-        },
-        pendingAnnotations: {
-          'temp-guid': {
-            id: 'temp-guid',
-            comment: 'pending annotation',
-            documentId
+          pendingAnnotations: {
+            'temp-guid': {
+              id: 'temp-guid',
+              comment: 'pending annotation',
+              documentId
+            }
           }
         }
       };
