+++ client/test/node/reader/DecisionReviewer-test.js
@@ -18,9 +18,8 @@ import { formatDateStr } from '../../../app/util/DateUtil';
 
 import readerReducer from '../../../app/reader/reducer';
 import caseSelectReducer from '../../../app/reader/CaseSelect/CaseSelectReducer';
-import documentsReducer from '../../../app/reader/DocumentList/DocumentsReducer';
 import { annotationLayerReducer } from '../../../app/reader/AnnotationLayer/AnnotationLayerReducer';
-
+import documentsReducer from '../../../app/reader/DocumentList/DocumentsReducer';
 import PdfJsStub, { PAGE_WIDTH, PAGE_HEIGHT } from '../../helpers/PdfJsStub';
 import { onReceiveDocs } from '../../../app/reader/LoadingScreen/LoadingScreenActions';
 import { onReceiveAnnotations } from '../../../app/reader/AnnotationLayer/AnnotationActions';
